+++ kaleki-commiter/src/main/java/ru/mirea/kalekicommiter/repositories/PositionRepository.java
@@ -7,7 +7,7 @@ import ru.mirea.kalekicommiter.entitys.Position;
 import java.util.Optional;
 
 @Repository
-public interface PositionRepository extends JpaRepository<Position, Integer> {
-    Optional<Position> findById(int position_id);
+public interface PositionRepository extends JpaRepository<Position, Long> {
+    Optional<Position> findById(long position_id);
     Optional<Position> findByName(String name);
 }
