+++ kaleki-commiter/src/main/java/ru/mirea/kalekicommiter/repositories/ProjectRepository.java
@@ -7,6 +7,6 @@ import ru.mirea.kalekicommiter.entitys.Project;
 import java.util.Optional;
 
 @Repository
-public interface ProjectRepository extends JpaRepository<Project, Integer> {
-    Optional<Project> findById(int project_id);
+public interface ProjectRepository extends JpaRepository<Project, Long> {
+    Optional<Project> findById(long project_id);
 }
